Front API application in Docker
