+++ .gitignore
@@ -1,4 +1,3 @@
-.env/
 qua.sublime*
 __pycache__/
 *.py[cod]
