+++ api/app/qua/api/exceptions.py
@@ -21,6 +21,7 @@ class ExitDecoratorError(QuaException):
 
 
 def custom_api_exception_handler(exc, context):
+
     response = exception_handler(exc, context)
 
     if response is not None and isinstance(response.data, dict):
