+++ api/app/qua/api/tests/test_auth.py
