+++ api/app/qua/api/utils/shortcuts.py
@@ -2,6 +2,7 @@ from rest_framework.exceptions import NotFound
 
 
 def get_object(cls, pk):
+
     try:
         return cls.objects.get(pk=pk)
     except cls.DoesNotExist:
