+++ api/app/qua/utils.py
@@ -1,8 +1,7 @@
-import logging
 from hashlib import sha256
-from urllib.parse import urlparse
 from importlib import import_module
-from datetime import datetime
+from urllib.parse import urlparse
+import logging
 
 from django.conf import settings
