+++ api/app/requirements.txt
@@ -1,0 +1,13 @@
+appdirs==1.4.3
+Django==1.11
+djangorestframework==3.6.2
+djangorestframework-jwt==1.10.0
+mistune==0.7.4
+packaging==16.8
+psycopg2==2.7.1
+PyJWT==1.5.0
+pyparsing==2.2.0
+pytz==2017.2
+redis==2.10.5
+requests==2.13.0
+six==1.10.0
