+++ api/app/wsgi.ini
@@ -1,15 +1,15 @@
 [uwsgi]
-chdir = /var/www
+chdir = /var/www/app
 env = DJANGO_SETTINGS_MODULE=qua.settings
 module = qua.wsgi:application
 
+plugins=python3
 master = true
 processes = 5
 
-uid = 1001
-gid = 1001
-socket = 127.0.0.1:9000
+pidfile = /var/run/qua-api.pid
+socket = /var/run/qua-api.sock
+chmod-socket = 666
 vacuum=True
 
 die-on-term = true
-logto = /var/log/qua/uwsgi.log
