+++ api/conf/.gitignore
@@ -1,0 +1,2 @@
+django_key
+db_password
