+++ api/conf/app.conf
@@ -1,0 +1,13 @@
+server {
+    server_name _;
+    listen 80;
+
+    location /static/ {
+        root /var/lib/qua/data;
+    }
+
+    location / {
+        include uwsgi_params;
+        uwsgi_pass unix:/var/run/qua-api.sock;
+    }
+}
