+++ api/conf/supervisord.conf
@@ -1,0 +1,8 @@
+[supervisord]
+nodaemon=true
+
+[program:nginx]
+command=/usr/sbin/nginx
+
+[program:uwsgi]
+command=/usr/bin/uwsgi --ini  /var/www/app/wsgi.ini
