--- api/requirements.txt
@@ -1,48 +1,0 @@
-alabaster==0.7.9
-amqp==2.1.4
-argcomplete==1.8.2
-Babel==2.3.4
-beautifulsoup4==4.5.3
-billiard==3.5.0.2
-bs4==0.0.1
-celery==4.0.2
-chardet==2.3.0
-DAWG==0.7.8
-DAWG-Python==0.7.2
-Django==1.10.5
-django-cors-headers==2.0.0
-djangorestframework==3.5.3
-djangorestframework-jwt==1.9.0
-docopt==0.6.2
-docutils==0.13.1
-docx2txt==0.6
-EbookLib==0.15
-elasticsearch==5.1.0
-imagesize==0.7.1
-Jinja2==2.9.4
-kombu==4.0.2
-lxml==3.7.2
-Markdown==2.6.7
-MarkupSafe==0.23
-mistune==0.7.3
-netaddr==0.7.19
-nltk==3.2.2
-olefile==0.44
-Pillow==4.0.0
-pkg-resources==0.0.0
-psycopg2==2.6.2
-Pygments==2.1.3
-PyJWT==1.4.2
-python-pptx==0.6.2
-pytz==2016.10
-redis==2.10.5
-requests==2.12.4
-six==1.10.0
-snowballstemmer==1.2.1
-SpeechRecognition==3.6.0
-Sphinx==1.5.1
-textract==1.5.0
-urllib3==1.20
-vine==1.1.3
-xlrd==1.0.0
-XlsxWriter==0.9.6
